[refactor] split re_ui_example into multiple files
